requirements.txt
oracledb==2.1.2
psycopg2-binary==2.9.9
PyMySQL==1.1.0
SQLAlchemy==2.0.28
pandas==2.2.1
python-dotenv==1.0.1
tqdm==4.66.2
